[MOB-7613] Remove extra height from iframe + refactor (#302)
* [MOB-7613] Unset iframe body margin if one is not explicitly already set

* [MOB-7613] Detect img tags alongside image url paths

* [MOB-7613] Set iframe height on iframe load

* [MOB-7613] Add jwt generator url to .env.example

* [MOB-7613] Cleaup utils

* [MOB-7613] Move caching methods to separate file

* [MOB-7613] Update usages of cache

* [MOB-7613] Clean up comments in cache.ts

* [MOB-7613] Remove added return

* [MOB-7613] Clean up iframe width/height setter

* [MOB-7613] Add comment for consume variable

* [MOB-7613] Add env convenience variable for react sample app

* [MOB-7613] Suppress console.warn lint warnings

* [MOB-7613] Clean up types and remove template literals

* [MOB-7613] Fix tests

* [MOB-7613] Fix tests

* [MOB-7613] Put localhost url as fallback for jwt generator
