+++ .env.example
@@ -2,10 +2,11 @@
 # called .env and add these values to it and change them appropriately.
 # Remember to uncomment the variables!
 
-# Only set BASE_URL if developing locally, as it will take precedence over the production api urls
+# Only set BASE_URL if developing locally, as it will take precedence over the production api urls.
 # BASE_URL="https://api.iterable.com/api"
 
+# Set this to false to prevent messages from being consumed to fetch the same message(s) when testing changes locally.
 # ENABLE_INAPP_CONSUME=false
 
-# toggle this to true if you would need to hit our EU APIs
+# Toggle this to true if you would need to hit our EU APIs.
 # IS_EU_ITERABLE_SERVICE=false
