+++ react-example/.env.example
@@ -1,5 +1,10 @@
 # To make requests from this example app make sure you first create an .env file
 # and add the API key and JWT Secret to it like so (and uncomment the keys):
-
 # API_KEY=1234
 # JWT_SECRET=1234
+
+# You can set the URL for the JWT generator here if needed
+# JWT_GENERATOR=http://localhost:5000/generate
+
+# Convenience variable to automatically set the login email during testing.
+# LOGIN_EMAIL=email@email.com
