+++ src/inapp/request.ts
@@ -1,6 +1,7 @@
 import { delMany, entries } from 'idb-keyval';
 import { GETMESSAGES_PATH, SDK_VERSION, WEB_PLATFORM } from 'src/constants';
 import { baseIterableRequest } from 'src/request';
+import { addNewMessagesToCache, getCachedMessagesToDelete } from './cache';
 import schema from './inapp.schema';
 import {
   CachedMessage,
@@ -8,7 +9,6 @@ import {
   InAppMessageResponse,
   InAppMessagesRequestParams
 } from './types';
-import { addNewMessagesToCache, getCachedMessagesToDelete } from './utils';
 
 type RequestInAppMessagesProps = {
   latestCachedMessageId?: string;
@@ -107,7 +107,8 @@ export const requestMessages = async ({ payload }: RequestMessagesProps) => {
     );
     try {
       await delMany(cachedMessagesToDelete);
-    } catch (err: any) {
+    } catch (err: any /* eslint-disable-line @typescript-eslint/no-explicit-any */) {
+      // eslint-disable-next-line no-console
       console.warn(
         'Error deleting messages from the browser cache',
         err?.response?.data?.clientErrors ?? err
@@ -124,7 +125,8 @@ export const requestMessages = async ({ payload }: RequestMessagesProps) => {
         inAppMessages: allMessages
       }
     };
-  } catch (err: any) {
+  } catch (err: any /* eslint-disable-line @typescript-eslint/no-explicit-any */) {
+    // eslint-disable-next-line no-console
     console.warn(
       'Error requesting in-app messages',
       err?.response?.data?.clientErrors ?? err
